+++ CHANGELOG.md
@@ -30,4 +30,4 @@ CHANGELOG
 * added default snapshot name (RunID_Pos) when used from pyscript, e.g. 12345-01A_3
 * added note, last_modified to irrad_ProductionTable
 * updated editing default_fits
-* added snapshot view to recall. currently only a list of snapshots not the actual images
+* added snapshot view to recall